refactor(core): share institution analysis; MCP uses evidence-based logic
Extract the web app's evidence-based institution analysis into
@nexus-score/core (institution/{types,analyze,index}.ts) so the web app
and the MCP server run one implementation that can't re-diverge.

- core: new institution module exported from the package index
- web: lib/publisher-map.ts and lib/analyze-institution.ts become thin
  re-export shims over @nexus-score/core (no import-site churn)
- mcp-server: analyze_institution now calls core analyzeInstitution and
  serializes the InstitutionReport; search_institutions uses core
  searchInstitutions. Removed the old PUBLISHER_MAP, host_organization_
  lineage grouping, and getPublisherCoverage aggregate-projection path
  that made the MCP output diverge from (and be less truthful than) the
  web app. Added an optional window_days arg (default 90).

core, web, and mcp-server all build clean.

Closes #10
